+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 - Initial boilerplate.
+- Add deploy script